+++ editor/env.ts
@@ -9,6 +9,39 @@ export namespace Env {
       : "http://localhost:3000";
   }
 
+  export namespace storage {
+    /**
+     * public, temporary file uploader to playground bucket
+     * for internal dev or public tmp playgrounds
+     *
+     * @public true
+     */
+    export const BUCKET_DUMMY = "dummy";
+
+    /**
+     * form media files (not response uploads)
+     *
+     * @public true
+     */
+    export const BUCKET_GRIDA_FORMS = "grida-forms";
+
+    /**
+     * private asset files
+     *
+     * @private
+     */
+    export const BUCKET_ASSETS = "assets";
+
+    /**
+     * public asset files
+     *
+     * for user uploaded contents in public cms
+     *
+     * @public
+     */
+    export const BUCKET_ASSETS_PUBLIC = "assets-public";
+  }
+
   /**
    * anything related to web & client side
    */
