+++ editor/scaffolds/platform/storage/index.ts
@@ -1,0 +1,2 @@
+export * from "./use-dummy-public-upload";
+export * from "./use-upload";
